requirements.txt
python-telegram-bot==20.7
xrpl-py
websockets
